feat(js): rename DatasetEvents as AssetEvents
